+++ src/main/java/onlyoffice/integration/sdk/managers/JwtManagerImpl.java
@@ -1,0 +1,43 @@
+/**
+ *
+ * (c) Copyright Ascensio System SIA 2023
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package onlyoffice.integration.sdk.managers;
+
+import com.onlyoffice.manager.security.DefaultJwtManager;
+import com.onlyoffice.manager.security.JwtManager;
+import com.onlyoffice.manager.settings.SettingsManager;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+
+@Component(
+    service = JwtManager.class
+)
+public class JwtManagerImpl extends DefaultJwtManager {
+
+    public JwtManagerImpl() {
+        super(null);
+    }
+
+    @Reference(service = SettingsManager.class, unbind = "-")
+    public void setSettingsManager(
+            SettingsManager settingsManager) {
+        super.setSettingsManager(settingsManager);
+    }
+}